+++ lib/mix/tasks/learn.ex
@@ -6,7 +6,8 @@ defmodule Mix.Tasks.Learn do
   @shortdoc "Learn Elixir through meditation"
 
   def run(_) do
-    Koans.start
-    Koans.load
+    Application.ensure_all_started(:koans)
+    System.at_exit(fn 0 -> Koans.Runner.run end)
+    Koans.Lessons.load
   end
 end