+++ test/lessons_test.exs
@@ -2,7 +2,7 @@ defmodule LessonsTest do
   use ExUnit.Case
 
   test "lessons compile" do
-    Koans.load
+    Koans.Lessons.load
     # no assertions needed here, compilation failure would cause a test failure
   end
 end